OpenWeatherMap/requirements.txt
affine==2.4.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contextily==1.5.2
contourpy==1.1.1
cramjam==2.8.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fastparquet==2024.2.0
fiona==1.9.6
fonttools==4.50.0
fsspec==2024.3.1
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.1
idna==3.6
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
jupyter-client==8.6.1
jupyter-core==5.7.2
kiwisolver==1.4.5
lazy-loader==0.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mercantile==1.2.1
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
pyparsing==3.1.2
pyproj==3.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
stack-data==0.6.3
tifffile==2023.7.10
tornado==6.4
traitlets==5.14.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xyzservices==2023.10.1
zipp==3.18.1
